Added a lot of missing Javadoc and improved many method implementations.
Used signpost's baseURI implementation to make that method more robust
and correctly follow the spec. Modified params to make skipping null
values more performant as well as signing.

Modified sign to throw an exception when a MAC cannot be generated as
that is an exception situation that the caller must be aware of.